+++ src/BuildingBlocks/Common/Contracts/Events/Payment/PaymentProcessedEvent.cs
@@ -1,0 +1,7 @@
+﻿namespace Common.Contracts.Events.Payment;
+
+public record PaymentProcessedEvent
+{
+    public Guid PaymentRequestId { get; init; }
+    public string TransactionId { get; init; }
+}