+++ src/Services/PaymentGateway/PaymentGateway.API/Extensions/ApplicationBuilderExtensions.cs
@@ -1,0 +1,21 @@
+﻿using Prometheus;
+using Serilog;
+
+namespace PaymentGateway.API.Extensions;
+
+public static class ApplicationBuilderExtensions
+{
+    public static WebApplication UseCustomMiddlewares(this WebApplication app)
+    {
+        app.UseInfrastructureMiddlewares();
+        return app;
+    }
+    
+    private static WebApplication UseInfrastructureMiddlewares(this WebApplication app)
+    {
+        app.UseHttpMetrics();
+        app.MapMetrics();
+        app.UseSerilogRequestLogging();
+        return app;
+    }
+}