+++ src/Services/PaymentOrchestration/PaymentOrchestration.Application/StateMachines/PaymentRequestSagaState.cs
@@ -9,4 +9,6 @@ public class PaymentRequestSagaState : SagaStateMachineInstance
 
     public decimal Amount { get; set; }
     public string Currency { get; set; }
+    public string TokenizedCardNumber { get; set; } 
+    public string CardHolderName { get; set; } 
 }