+++ src/Services/PaymentOrchestration/PaymentOrchestration.Application/StateMachines/PaymentRequestSagaStateMachine.cs
@@ -3,6 +3,7 @@ using Common.Contracts.Events.Fraud;
 using Common.Contracts.Events.Payment;
 using MassTransit;
 using Microsoft.Extensions.Logging;
+using PaymentOrchestration.Application.StateMachines.Activities;
 
 namespace PaymentOrchestration.Application.StateMachines;
 
@@ -10,12 +11,16 @@ public class PaymentRequestSagaStateMachine : MassTransitStateMachine<PaymentReq
 {
     private readonly ILogger<PaymentRequestSagaStateMachine> _logger;
     public State Submitted { get; private set; }
+    public State FraudCheckPassed { get; private set; } 
+    public State PaymentProcessing { get; private set; } //
     public State Completed { get; private set; }
     public State Failed { get; private set; }
     
     public Event<PaymentRequestSubmittedEvent> PaymentRequestSubmitted { get; private set; }
-    public Event<FraudCheckPassedEvent> FraudCheckPassed { get; private set; }
+    public Event<FraudCheckPassedEvent> FraudCheckPassedEvent { get; private set; }
     public Event<FraudCheckFailedEvent> FraudCheckFailed { get; private set; }
+    public Event<PaymentProcessedEvent> PaymentProcessedEvent { get; private set; }
+    public Event<PaymentProcessingFailedEvent> PaymentProcessingFailedEvent { get; private set; }
 
     public PaymentRequestSagaStateMachine(ILogger<PaymentRequestSagaStateMachine> logger)
     {
@@ -24,14 +29,20 @@ public class PaymentRequestSagaStateMachine : MassTransitStateMachine<PaymentReq
         InstanceState(x => x.CurrentState);
         
         Event(() => PaymentRequestSubmitted, x => x.CorrelateById(context => context.Message.PaymentRequestId));
-        Event(() => FraudCheckPassed, x => x.CorrelateById(context => context.Message.PaymentRequestId));
+        Event(() => FraudCheckPassedEvent, x => x.CorrelateById(context => context.Message.PaymentRequestId));
         Event(() => FraudCheckFailed, x => x.CorrelateById(context => context.Message.PaymentRequestId));
+        Event(() => PaymentProcessedEvent, x => x.CorrelateById(context => context.Message.PaymentRequestId));
+        Event(() => PaymentProcessingFailedEvent, x => x.CorrelateById(context => context.Message.PaymentRequestId));
+        
+        
         Initially(
             When(PaymentRequestSubmitted)
                 .Then(context =>
                 {
                     context.Saga.Amount = context.Message.Amount;
                     context.Saga.Currency = context.Message.Currency;
+                    context.Saga.CardHolderName = context.Message.CardHolderName;
+                    context.Saga.TokenizedCardNumber = context.Message.TokenizedCardNumber;
                     _logger.LogWarning("SAGA Started for Payment ID: {PaymentId}", context.Message.PaymentRequestId);
                 }).TransitionTo(Submitted).Publish(context => new CheckFraudCommand
                 {
@@ -42,14 +53,26 @@ public class PaymentRequestSagaStateMachine : MassTransitStateMachine<PaymentReq
             );
         
         During(Submitted,
-            When(FraudCheckPassed)
+            When(FraudCheckPassedEvent)
                 .Then(context => _logger.LogInformation("Fraud check PASSED for Payment ID: {CorrelationId}", context.Saga.CorrelationId))
-                .TransitionTo(Completed)
-                .Finalize(),
+                .Activity(x=> x.OfType<RoutingAndBankDispatchingActivity>())
+                .TransitionTo(PaymentProcessing),
             When(FraudCheckFailed)
                 .Then(context =>  _logger.LogError("Fraud check FAILED for Payment ID: {CorrelationId}. Reason: {Reason}", context.Saga.CorrelationId, context.Message.Reason))
                 .TransitionTo(Failed)
                 .Finalize());
+
+        During(PaymentProcessing,
+            When(PaymentProcessedEvent)
+                .Then(context => _logger.LogInformation("Payment PROCESSED for Payment ID: {CorrelationId}. Bank Transaction ID: {TransactionId}", context.Saga.CorrelationId, context.Message.TransactionId))
+                .TransitionTo(Completed)
+                .Finalize(),
+
+            When(PaymentProcessingFailedEvent)
+                .Then(context => _logger.LogError("Payment processing FAILED for Payment ID: {CorrelationId}. Reason: {Reason}", context.Saga.CorrelationId, context.Message.Reason))
+                .TransitionTo(Failed)
+                .Finalize()
+        );
         
         SetCompletedWhenFinalized();